+++ pycatia/__init__.py
@@ -14,7 +14,6 @@ from .hybridshapefactory import HybridShapeFactory
 from .measurable import CATIAMeasurable
 from .part import Part
 from .product import Product
-from .reference import create_reference
 from .workbenches import create_spa_workbench
 
 
@@ -22,7 +21,7 @@ __author__ = 'Paul Bourne'
 __author_email = 'evereux@gmail.com'
 __description__ = 'A python module to access the CATIA Measurable object.'
 __name__ = "pycatia"
-__version__ = "0.1.2"
+__version__ = "0.1.3"
 __url__ = "https://github.com/evereux/pycatia"
 
 name = __name__