+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.1.3
+
+* Added in_work_object method to class Part().
+* Moved create_reference method to class Part(). It's proper home.
+* Added find_object_by_name method to class Part().
+
 ## 0.1.2
 
 * Added support to export files along with example.